Move backtrace to its own header
